Separate distance parsing
